Started GetMeeting()
I'm not sure anymore that I need to have to meeting list one for past
and future. It is convenient to just iterate over one or the other but
the contract is making me iterate over both a the same time all the
time. My need to refactor. However, the editor is expecting it so I
would have to refactor there too :(